chore(pattern): rename ContentBlockMixedGroups to ContentBlockMixed
