+++ packages/react/src/patterns/blocks/ContentBlockMixed/__stories__/index.scss
