+++ packages/react/src/patterns/blocks/ContentBlockMixed/index.js
@@ -5,4 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-export { default as ContentBlockMixedGroups } from './ContentBlockMixedGroups';
+export { default as ContentBlockMixed } from './ContentBlockMixed';
